Move `WriteThroughCursor` internals to `alloc::io`
